Odbc (#20)
* Switched from MySQLConnector to an ODBC Client

* Remove unsigned and changed tinyint to smallint. Still WIP.

* Filled up the rest of odbc call statements

* Updated Admin app for ODBC

* Added default back in when updating column information

* Switched admin app to use existing stored procedures

* Added a tidbit to the readme

* Cleaned up comment and using

* took care of a few unsigned int16

* Removed INI reading from web project

* Replaced missed tinyint with smallint

+++ EnvelopeWeb/Models.fs
@@ -6,7 +6,7 @@ type ErrorViewModel =
     member this.ShowRequestId = not (System.String.IsNullOrEmpty(this.RequestId))
 
 [<CLIMutable>]
-type Envelope = { Number: uint16; Name: string; Amount: double }
+type Envelope = { Number: int16; Name: string; Amount: double }
 
 [<CLIMutable>]
 type EnvelopeName = { EnvelopeName: string }
@@ -17,22 +17,22 @@ type Login = { Email: string; Password: string; }
 type LoginSignUpErrors = { EmailError: string; PasswordError: string; ConfirmPasswordError: string }
 
 [<CLIMutable>]
-type PasswordConfig = { MiB: byte; Iterations: byte; DegreeOfParallism: byte; Salt: byte array }
+type PasswordConfig = { MiB: int16; Iterations: int16; DegreeOfParallism: int16; Salt: byte array }
 
 [<CLIMutable>]
 type SignUp = { Email: string; Password: string; ConfirmPassword: string; }
 
 [<CLIMutable>]
-type Transaction = { TransactionNumber: uint; Amount: double; Date: System.DateTime; Note: string; }
+type Transaction = { TransactionNumber: int; Amount: double; Date: System.DateTime; Note: string; }
 
 [<CLIMutable>]
-type Transfer = { DestinationNumber: uint16; Amount: double; }
+type Transfer = { DestinationNumber: int16; Amount: double; }
 
 [<CLIMutable>]
-type Sel_Transactions_Result = { NumberOfAllTransactions: uint; Transactions: Transaction list}
+type Sel_Transactions_Result = { NumberOfAllTransactions: int; Transactions: Transaction list}
 
 [<CLIMutable>]
-type User = { Id: uint; PasswordHash: byte array; LockoutExpiry: System.DateTime; PasswordConfig: PasswordConfig; }
+type User = { Id: int; PasswordHash: byte array; LockoutExpiry: System.DateTime; PasswordConfig: PasswordConfig; }
 
 [<CLIMutable>]
 type YesNo = { YesNo: string }
